changed helper function to a separate file and reused them in worker thread
