+++ backend/src/modules/question/controllers/QuestionController.ts
@@ -202,7 +202,7 @@ export class QuestionController {
           payload,
           isOutreachQuestion
         );
-        setImmediate(() => startBackgroundProcessing(insertedIds, isRequiredAiInitialAnswer));
+        setImmediate(() => startBackgroundProcessing(insertedIds, isRequiredAiInitialAnswer, isOutreachQuestion));
         return {
           message: `✅ Successfully uploaded ${insertedIds.length} question(s). The expert allocation process has been initiated.${isRequiredAiInitialAnswer
               ? " AI-generated initial answers will be included for each question."
